Rename `.binance.schemas` -> `.venues`
